test: move VerifyReturnTypeTests into pyspark.tests.test_worker
